+++ api/kyverno/v1/rule_types.go
@@ -81,8 +81,9 @@ func (r *Rule) HasGenerate() bool {
 	return !reflect.DeepEqual(r.Generation, Generation{})
 }
 
-func (r *Rule) MutatingExisting() bool {
-	return r.Mutation.MutateExisting
+// IsMutatingExisting checks if the mutate rule applies to existing resources
+func (r *Rule) IsMutateExisting() bool {
+	return r.Mutation.Targets != nil
 }
 
 func (r *Rule) GetAnyAllConditions() apiextensions.JSON {
