+++ api/kyverno/v1beta1/updaterequest_types.go
@@ -65,8 +65,19 @@ type UpdateRequest struct {
 	Status UpdateRequestStatus `json:"status,omitempty"`
 }
 
+type RequestType string
+
+const (
+	Mutate   RequestType = "mutate"
+	Generate RequestType = "generate"
+)
+
 // UpdateRequestSpec stores the request specification.
 type UpdateRequestSpec struct {
+	// Type represents request type for background processing
+	// +kubebuilder:validation:Enum=mutate;generate
+	Type RequestType `json:"requestType,omitempty" yaml:"requestType,omitempty"`
+
 	// Specifies the name of the policy.
 	Policy string `json:"policy" yaml:"policy"`
